initial work on incorporating REPL classes.
some cleanup and re-org. untested

basic example working

loading classes defined in REPL still doesn't work

started java reorg

Reorganized code. Everything compiles

Finished implementation. Now need docker container and lots of debugging.

start work on docker container

fixed bug in docker file

Fixed some bugs

Clipper local deploy seems to be working

Fixed some bugs

Deploy is working but there's something wrong with the container

Remote deploy is working!

Spark scala deploy is working both locally and remotely

Formatting

started to add tests
