--- containers/java/clipper-java-container/src/main/java/clipper/container/app/NoOpStringModel.java
@@ -1,16 +1,0 @@
-package clipper.container.app;
-
-import clipper.container.app.data.DataType;
-import clipper.container.app.data.SerializableString;
-
-public class NoOpStringModel extends Model<SerializableString> {
-  NoOpStringModel(String name, int version) {
-    super(name, version, DataType.Strings);
-  }
-
-  @Override
-  public SerializableString predict(SerializableString inputVector) {
-    String jsonContent = String.format("{ \"data_size\": %d }", inputVector.getData().length());
-    return new SerializableString(jsonContent);
-  }
-}
