--- containers/java/clipper-java-container/src/main/java/clipper/container/app/data/DataBuffer.java
@@ -1,11 +1,0 @@
-package clipper.container.app.data;
-
-import java.nio.ByteBuffer;
-
-public abstract class DataBuffer<T> {
-  abstract void init(ByteBuffer inputBuffer);
-
-  abstract T get(int size);
-
-  abstract T getAll();
-}
