--- containers/java/clipper-java-container/src/main/java/clipper/container/app/data/DataVector.java
@@ -1,13 +1,0 @@
-package clipper.container.app.data;
-
-public abstract class DataVector<T> {
-  T data;
-
-  DataVector(T data) {
-    this.data = data;
-  }
-
-  public T getData() {
-    return data;
-  }
-}
