+++ containers/jvm/clipper-container/README.md
@@ -1,0 +1,3 @@
+Maven module that is intended to be included as a dependency in Java
+or Scala model training programs. It contains the Clipper model
+container interface and associated classes.
