+++ containers/jvm/clipper-container/src/main/java/ai/clipper/container/data/DataVector.java
@@ -1,0 +1,13 @@
+package ai.clipper.container.data;
+
+public abstract class DataVector<T> {
+  T data;
+
+  DataVector(T data) {
+    this.data = data;
+  }
+
+  public T getData() {
+    return data;
+  }
+}
